(chore) test with mobile css
