+++ src/components/tools/routing/drivers/driver-form.tsx
@@ -157,9 +157,9 @@ export const DriverForm: FC<TDriverForm> = ({ handleOnOpenChange }) => {
     <Form {...form}>
       <form
         onSubmit={(e) => void form.handleSubmit(onSubmit)(e)}
-        className=" flex  h-full flex-grow flex-col space-y-8 lg:justify-between"
+        className=" flex h-full flex-col justify-between space-y-8  lg:flex-grow"
       >
-        <ScrollArea className="relative flex h-full max-h-[75vh] w-full flex-1 max-md:max-h-[60vh]">
+        <ScrollArea className="relative flex h-full  w-full flex-1  max-md:max-h-[60vh]">
           <div className=" flex  w-full  flex-col space-y-4">
             <FormField
               control={form.control}
