+++ src/components/tools/routing/stops/stop-form.tsx
@@ -160,9 +160,9 @@ export const StopForm: FC<TStopForm> = ({ handleOnOpenChange }) => {
     <Form {...form}>
       <form
         onSubmit={(e) => void form.handleSubmit(onSubmit)(e)}
-        className=" flex h-full flex-col justify-between space-y-8 bg-yellow-500 lg:flex-grow"
+        className=" flex h-full flex-col justify-between space-y-8  lg:flex-grow"
       >
-        <ScrollArea className="relative flex h-full  w-full flex-1 bg-green-500 max-md:max-h-[60vh]">
+        <ScrollArea className="relative flex h-full  w-full flex-1  max-md:max-h-[60vh]">
           <div className=" w-full space-y-4">
             <FormField
               control={form.control}
@@ -419,7 +419,7 @@ export const StopForm: FC<TStopForm> = ({ handleOnOpenChange }) => {
             </div>
           </div>
         </ScrollArea>
-        <div className="mt-auto flex gap-4 bg-blue-500">
+        <div className="mt-auto flex gap-4 ">
           <Button
             type="button"
             className="flex gap-4"
